Refactor: 채팅방 서비스 Command, Query로 분리
