+++ src/main/java/com/gdg/z_meet/domain/chat/controller/ChatWebSocketController.java
@@ -2,7 +2,6 @@ package com.gdg.z_meet.domain.chat.controller;
 
 
 import com.gdg.z_meet.domain.chat.dto.ChatMessage;
-import com.gdg.z_meet.domain.chat.service.ChatRoomService;
 import com.gdg.z_meet.domain.chat.service.ChatService;
 import com.gdg.z_meet.domain.chat.service.MessageService;
 import com.gdg.z_meet.domain.user.entity.User;
@@ -28,13 +27,9 @@ import java.util.UUID;
 @RequestMapping("/api/chat")
 public class ChatWebSocketController {
 
-    private final SimpMessagingTemplate messagingTemplate;
-    private final ChatRoomService chatRoomService;
-    private final MessageService messageService;
     private final JwtUtil jwtUtil;
     private final UserRepository userRepository;
-    private final RedisTemplate<String, Object> redisTemplate;
-    private final ChatService chatService; // ChatService 추가
+    private final ChatService chatService;
 
     @MessageMapping("/{roomId}")
     public void sendMessage(@DestinationVariable Long roomId, @Payload ChatMessage chatMessage, @Header("Authorization") String token) {
